moved function to correct utils
