+++ packages/zendesk-adapter/src/change_validators/duplicate_id_field_values.ts
@@ -23,8 +23,8 @@ import _ from 'lodash'
 import { elements as elementUtils } from '@salto-io/adapter-components'
 import { logger } from '@salto-io/logging'
 import { values as lowerDashValues } from '@salto-io/lowerdash'
+import { getInstancesFromElementSource } from '@salto-io/adapter-utils'
 import { ZendeskApiConfig } from '../config'
-import { getInstancesFromElementSource } from '../filters/article/utils'
 
 const { generateInstanceNameFromConfig } = elementUtils
 const { isDefined } = lowerDashValues
