+++ packages/zendesk-adapter/src/filters/article/utils.ts
@@ -31,7 +31,7 @@ import {
   isReferenceExpression,
   isStaticFile,
   isTemplateExpression,
-  ObjectType, ReadOnlyElementsSource,
+  ObjectType,
   ReferenceExpression,
   StaticFile,
   Values,
@@ -253,11 +253,3 @@ export const updateArticleTranslationBody = async ({
       })
     })
 }
-
-export const getInstancesFromElementSource = async (elementSource: ReadOnlyElementsSource, typeName: string)
-  : Promise<InstanceElement[]> =>
-  awu(await elementSource.list())
-    .filter(id => id.typeName === typeName)
-    .filter(id => id.idType === 'instance')
-    .map(id => elementSource.get(id))
-    .toArray()
